pyproject.toml
[tool.poetry]
name = "orchestra"
version = "0.1.0"
description = "Multi-agent orchestration system built on claude-agent-sdk"
authors = ["yanhs"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
claude-agent-sdk = ">=0.1.39"
click = ">=8.1"
pyyaml = ">=6.0"
rich = ">=13.0"
fastapi = ">=0.115"
uvicorn = {version = ">=0.34", extras = ["standard"]}

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"
pytest-asyncio = ">=0.23"
pytest-mock = ">=3.14"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.poetry.scripts]
orchestra = "src.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
